Day085/py-desktop-app/requirements.txt
altair==5.1.2
annotated-types==0.6.0
ansi2html==1.8.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==2.6.6
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.9.3
bleach==6.1.0
blinker==1.6.2
blis==0.7.11
Bootstrap-Flask==2.2.0
boto3==1.18.21
botocore==1.21.21
cached-property==1.5.2
cachetools==5.3.2
catalogue==2.0.10
certifi==2021.5.30
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.3
click==8.1.6
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.4
confection==0.1.3
contourpy==1.0.7
cryptography==41.0.5
cycler==0.11.0
cymem==2.0.8
dash==2.14.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dnspython==2.4.0
docx==0.2.4
docx2txt==0.8
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.12.2
Flask==2.2.5
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.39.3
fpdf==1.7.2
fqdn==1.5.1
ftfy==6.1.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==1.1.1
h11==0.14.0
httpcore==0.17.3
idna==3.2
ijson==3.1.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
ipykernel==6.26.0
ipython==8.17.2
isodate==0.6.0
isoduration==20.11.0
isort==5.9.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
json5==0.9.14
jsonlines==2.0.0
jsonpointer==2.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.2
kiwisolver==1.4.4
langcodes==3.3.0
lazy-object-proxy==1.6.0
linear-tsv==1.1.0
lxml==4.6.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
mistune==3.0.2
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.21.1
openpyxl==3.0.7
overrides==7.4.0
packaging==23.2
pandas==1.3.0
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20221105
Pillow==9.5.0
platformdirs==3.11.0
plotly==5.17.0
preshed==3.0.9
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pylint==2.9.6
pyparsing==3.0.9
pyperclip==1.8.2
pyresparser==1.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==0.19.0
python-json-logger==2.0.7
pytz==2021.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
regex==2023.3.23
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.13.1
s3transfer==0.5.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
selenium==3.141.0
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post1
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spotipy==2.18.0
SQLAlchemy==1.4.22
srsly==2.4.8
stack-data==0.6.3
streamlit==1.28.1
tableschema==1.20.2
tabulator==1.53.5
tenacity==8.2.3
terminado==0.18.0
thinc==8.2.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
unicodecsv==0.14.1
uri-template==1.3.0
urllib3==1.26.6
validators==0.22.0
virtualenv==20.24.2
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.9
weasel==0.3.3
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.2.3
wrapt==1.12.1
WTForms==3.0.1
xlrd==2.0.1
zipp==3.17.0
